Work with long

+++ vanilla/src/main/java/vanillajavaexamples/bitwise/DateBitwise.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class DateBitwise {
 
-  public static int toFlags(List<LocalDate> dates) {
+  public static long toFlags(List<LocalDate> dates) {
     return dates
         .stream()
         .map(it -> toFlag(dates.get(0), it))
@@ -14,25 +14,25 @@ public class DateBitwise {
         .get();
   }
 
-  public static int toFlag(LocalDate first, LocalDate date) {
+  public static long toFlag(LocalDate first, LocalDate date) {
     final var days = ChronoUnit.DAYS.between(first, date);
-    return Math.max(1, 1 << days);
+    return Math.max(1L, 1L << days);
 //    return Integer.parseInt(date.toString().replace("-", ""));
   }
 
-  public static boolean hasFlag(int flags, int flag) {
+  public static boolean hasFlag(long flags, long flag) {
     return (flags & flag) == flag;
   }
 
-  public static int removeFlag(int flags, int flag) {
+  public static long removeFlag(long flags, long flag) {
     return flags & (~flag);
   }
 
-  public static boolean hasDate(int flags, LocalDate first, LocalDate d) {
+  public static boolean hasDate(long flags, LocalDate first, LocalDate d) {
     return hasFlag(flags, toFlag(first, d));
   }
 
-  public static int removeDate(int flags, LocalDate first, LocalDate d) {
+  public static long removeDate(long flags, LocalDate first, LocalDate d) {
     return removeFlag(flags, toFlag(first, d));
   }
